cleanup: move 'maxGas' constants to 'contracts' module
Also, moves DebitFee and CreditFee to 'contracts' module
